feat: Make our custom time picker dialog reusable for futur snooze feature
I've split in two classes the logic that's specific to the schedule dialog and the logic that's common and generic to a reusable time picker dialog
