+++ app/src/main/java/com/infomaniak/mail/ui/newMessage/NewMessageFragment.kt
@@ -224,7 +224,7 @@ class NewMessageFragment : Fragment() {
 
         getBackNavigationResult(OPEN_DATE_AND_TIME_SCHEDULE_DIALOG) { _: Boolean ->
             dateAndTimeScheduleDialog.show(
-                onSchedule = { timestamp ->
+                onDateSelected = { timestamp ->
                     localSettings.lastSelectedScheduleEpoch = timestamp
                     scheduleDraft(timestamp)
                 },
