+++ app/src/main/res/layout/dialog_select_date_and_time_for_scheduled_draft.xml
@@ -74,7 +74,7 @@
     </com.infomaniak.lib.core.views.EndIconTextInputLayout>
 
     <TextView
-        android:id="@+id/scheduleDateError"
+        android:id="@+id/errorMessage"
         style="@style/Body.Error"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
